Move `design/defaultTheme.js` down a level to work around hot reload
- Similarly to last commit, move `defaultTheme.js` down a level so that
   it works around the hot reload bug with imports at top level of packages
- (note this is currently duplicative with the theme in `tdesign`, and at some
    point `design` really needs to be converted to typescript and moved into
    `tdesign`)
